Created a README.md file
TodoList Project

My simple TodoList helps to manage tasks.
The user will be able to:

-Add tasks
-Remove tasks
-Change the completed and uncompleted status of a task

My TodoList with all its tasks and their status can also be displayed for the user, which I find to be important.

In order for my code to be well structured and organized I will have
several different files in order to to that.

Starting with the project I know that I will need at least three
different files which are:

- package.json
- tasks.csv
- todolist.js

Most propably I will later in the course of my project need to add or
restructure my files and code, but that is the fun part of it, learning
by doing.