Deleted Task class
I have deleted the Task class from the todolist.js file because I was
thinking it would be better to create a new file for it since I will
have to code another class too. In order for my code to be better
organised I have decided to do it this way.