Defined saveToCsv method
In order to save tasks to my csv file.
This method will convert the tasks to csv format and also save them in
the tasks.csv file. The good thing about it is that the current state of
the todo list will be saved for the next time I use the application.